backend/requirements.txt
alembic
annotated-types
anyio
async-timeout
asyncpg
Authlib
bcrypt
certifi
cffi
charset-normalizer
cleo
clerk-backend-api
click
clikit
colorama
crashtest
cryptography
dnspython
dotty-dict
ecdsa
email_validator
exceptiongroup
exceptionite
Faker
fastapi
h11
hashids
hfilesize
httpcore
httpx
hupper
idna
inflection
iniconfig
Jinja2
Mako
MarkupSafe
masonite-orm
mock==5.1.0
packaging==24.2
passlib==1.7.4
pastel==0.2.1
pendulum==2.1.2
phonenumbers==8.13.55
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT
pylev==1.4.0
pytest
pytest-asyncio
python-dateutil==2.9.0.post0
python-dotenv==0.15.0
python-jose==3.4.0
python-multipart==0.0.20
pytzdata==2020.1
requests==2.32.3
requests-file==2.1.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
stripe==11.6.0
tldextract==2.2.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchdog==4.0.0
Werkzeug==3.1.3
whitenoise==5.2.0
